+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - add `keep_fnames` option to preserve function names in compressed code
 - backwards incompatible changes to source map options
 - support for inline base64 encoded source maps
+- mangle property names
+- rename name mangle option to mangle_names
 
 ## 2.7.2 (26 August 2015)